third_party/WebKit/LayoutTests/inspector/device-orientation-success-expected.txt

CONSOLE MESSAGE: line 23: alpha: 20 beta: 30 gamma: 40

Running: setUp

Running: setOverride

Running: clearOverride
